Fix React hot reloading for IsLoading component.
